requirements.txt
pdfplumber
python-docx
reportlab
langdetect
